+ add logins / keys

+++ blog/views.py
@@ -4,24 +4,31 @@ from __future__ import unicode_literals
 from django.shortcuts import render, redirect
 from django.contrib.auth import authenticate, login, logout
 from django.contrib import messages
+from django.contrib.auth.decorators import login_required
 
 from .models import Todo
 
-# Create your views here.
+
 def blogIndex(request):
-    todos = Todo.objects.all()
-    context = {
-        'title': 'ToDos',
-        'todos': todos
-    }
-    print request.session.session_key
-    current_user = request.user
-    print current_user.id
-    return render(request, 'index.html', context)
+    #todos = Todo.objects.all()
+    if request.user.is_authenticated:
+        todos = Todo.objects.filter(owner=request.user.id)
+        context = {
+            'title': 'ToDos',
+            'todos': todos
+        }
+        print request.session.session_key
+        current_user = request.user
+        print current_user.id
+        return render(request, 'index.html', context)
+    else:
+        print 'nicht logged in'
+        return redirect('login')
 
 def blogHund(request):
     return render(request, 'index.html', {'title': 'Test', 'discription': 'dies ist ein test'})
 
+@login_required
 def details(request, todo_id):
     details = Todo.objects.get(pk=todo_id)
     context = {
@@ -32,30 +39,37 @@ def details(request, todo_id):
 
     return render(request, 'detail.html', context)
 
+@login_required
 def new(request):
     context = {
         'title': 'Add new'
     } 
     return render(request, 'new.html', context)
 
+@login_required
 def create_todo(request):
     todo = Todo()
     todo.title = request.GET['title']
     todo.body = request.GET['body']
+    todo.owner = request.user
     if todo.title and todo.body:
         todo.save()
     return redirect('index')
 
+@login_required
 def del_todo(request, todo_id):
     todo = Todo.objects.get(pk=todo_id)
-    todo.delete()
+    if request.user == todo.owner:
+        todo.delete()
     return redirect('index')
 
+@login_required
 def modify_todo(request, todo_id):
     todo = Todo.objects.get(pk=todo_id)
     todo.title = request.GET['title']
     todo.body = request.GET['body']
-    todo.save()
+    if request.user == todo.owner:
+        todo.save()
     return redirect('index')
 
 def log_in(request):
@@ -74,8 +88,9 @@ def log_on(request):
         return redirect('index')
     else:
         messages.add_message(request, messages.WARNING, "Wrong Password or Username!")
-        return redirect('index')
+        return redirect('login')
 
+@login_required
 def log_out(request):
     logout(request)
     return redirect('index')
